student and attendance with separate models

+++ Assignment2_SDLabApp/src/main/java/com/example/Assignment2_LabApp/controller/StudentController.java
@@ -1,14 +1,22 @@
 package com.example.Assignment2_LabApp.controller;
 
+import com.example.Assignment2_LabApp.apimodel.StudentRequestModel;
+import com.example.Assignment2_LabApp.apimodel.StudentResponseModel;
 import com.example.Assignment2_LabApp.model.Student;
 import com.example.Assignment2_LabApp.service.IStudentService;
+import com.example.Assignment2_LabApp.util.TokenGenerator;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static org.springframework.web.bind.annotation.RequestMethod.*;
 
@@ -19,34 +27,65 @@ public class StudentController {
     @Autowired
     private IStudentService studentService;
 
+    private ModelMapper modelMapper = new ModelMapper();
+
     @RequestMapping(method = GET)
-    public List<Student> getAllStudents(){
-        return studentService.getAllStudents();
+    public List<StudentResponseModel> getAllStudents(){
+        return studentService.getAllStudents().stream()
+                                              .map(stud -> modelMapper.map(stud, StudentResponseModel.class))
+                                              .collect(Collectors.toList());
     }
 
     @RequestMapping(method = GET, value = "/{id}")
-    public Student getUserById(@PathVariable int id){
-        return studentService.getStudentById(id);
+    public ResponseEntity getStudentById(@PathVariable int id){
+        Student stud = studentService.getStudentById(id);
+        if(stud != null)
+            return ResponseEntity.ok(modelMapper.map(stud, StudentResponseModel.class));
+        else
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Student not found.");
     }
 
     @RequestMapping(method = POST)
-    public void addStudent(@RequestBody Student student){
-        studentService.addStudent(student);
+    public ResponseEntity addStudent(@Validated @RequestBody StudentRequestModel student){
+        Student stud = modelMapper.map(student, Student.class);
+        stud.setIsTeacher(false);
+        System.out.println(student.getGroupName());
+        studentService.addStudent(stud);
+        return ResponseEntity.ok("New student added.");
     }
 
     @RequestMapping(method = PUT, value = "/{id}")
-    public void updateStudent(@RequestBody Student student){
-        //TODO check if Student exists
-        studentService.updateStudent(student);
+    public ResponseEntity updateStudent(@RequestBody StudentRequestModel student, int id){
+        Student oldStud = studentService.getStudentById(id);
+        if(oldStud != null) {
+            Student newStud= modelMapper.map(student, Student.class);
+            newStud.setId(id);
+            newStud.setIsTeacher(false);
+            newStud.setPassword(oldStud.getPassword());
+            studentService.updateStudent(newStud);
+            return ResponseEntity.ok("Student updated");
+        } else
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Student not found.");
     }
 
     @RequestMapping(method = DELETE, value = "/{id}")
-    public void deleteStudent(@PathVariable int id){
-        studentService.deleteStudent(id);
+    public ResponseEntity deleteStudent(@PathVariable int id){
+        if(studentService.getStudentById(id) != null){
+            studentService.deleteStudent(id);
+            return ResponseEntity.ok().body("Student deleted.");
+        } else
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Student not found.");
     }
 
-//    @RequestMapping(method = PUT, value = "/{id}/password")
-//    public void changePassword(@PathVariable id, ){
-//        studentService.changePassword(password);
-//    }
+    @RequestMapping(method = PUT, value = "/{id}/password")
+    public ResponseEntity changePassword(@PathVariable int id, @PathVariable String password){
+        Student stud = studentService.getStudentById(id);
+        if(stud == null)
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Student not found.");
+        else {
+            //TODO add authentication part
+            studentService.changePassword(stud, password);
+            return ResponseEntity.ok("Password changed.");
+        }
+    }
 }
